+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [[Bug]](https://www.notion.so/cere/If-a-player-took-20th-place-or-higher-then-21st-place-has-a-gift-icon-5bb7e12a9e4544a29f21bc607b3df324?pvs=4) Fix bug with rank
 - [[Bug]](https://www.notion.so/cere/A-rewarding-pop-up-appears-on-the-top-and-moves-the-layout-11a377734b264549aa70b75bab664333?pvs=4) Fix reward notification position on mobile
 - [[Bug]](https://www.notion.so/cere/Confetti-animation-on-Congratulations-screen-does-not-always-work-the-first-time-f88c1b30b9074077b70bd400f6e46a96?pvs=4) Preload static assets in advance to fix missing confetti animation
+- [[Bug]](https://www.notion.so/cere/After-interacting-with-Wallet-button-the-control-arrows-become-inactive-b4f58387f306481ead213af8cc7af364?pvs=4) Restore focus on game window after closing the wallet widget to fix keyboard interaction issue
 
 ### v0.2.0