+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Release Notes:
 
-### vNext
+### v0.3.0
 
 - [[Bug]](https://www.notion.so/cere/If-a-player-took-20th-place-or-higher-then-21st-place-has-a-gift-icon-5bb7e12a9e4544a29f21bc607b3df324?pvs=4) Fix bug with rank
 - [[Bug]](https://www.notion.so/cere/A-rewarding-pop-up-appears-on-the-top-and-moves-the-layout-11a377734b264549aa70b75bab664333?pvs=4) Fix reward notification position on mobile